Minor code style and comment changes
