+++ raiden_contracts/tests/fixtures/__init__.py
@@ -11,3 +11,4 @@ from .token_network_registry import *
 from .channel import *
 from .monitoring_service import *
 from .raiden_service_bundle_fixtures import *
+from .user_deposit import *
