+++ raiden_contracts/tests/test_user_deposit_contract.py
@@ -1,18 +1,4 @@
-import pytest
-
-from raiden_contracts.constants import CONTRACT_USER_DEPOSIT, CONTRACTS_VERSION
-
-
-@pytest.fixture
-def user_deposit_contract(
-    deploy_tester_contract,
-    custom_token,
-):
-    return deploy_tester_contract(
-        CONTRACT_USER_DEPOSIT,
-        {},
-        [custom_token.address],
-    )
+from raiden_contracts.constants import CONTRACTS_VERSION
 
 
 def test_deposit(
